Move mapper to Mappings.Common namespace
